🚚 Final pass at moving component files around
