+++ app/components/rad-card/body.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-radical/components/rad-card/body';
