--- app/components/rad-card/component.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-radical/components/rad-card/component';
