+++ app/components/rad-card/footer.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-radical/components/rad-card/footer';
