+++ app/components/rad-card/title.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-radical/components/rad-card/title';
