--- app/components/rad-card/title/component.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-radical/components/rad-card/title/component';
