--- app/components/rad-drawer/component.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-radical/components/rad-drawer/component';
