+++ app/components/rad-drawer/content.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-radical/components/rad-drawer/content';
