+++ app/components/rad-dropdown/content.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-radical/components/rad-dropdown/content';
