+++ app/components/rad-dropdown/menu-item.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-radical/components/rad-dropdown/menu-item';
