--- app/components/rad-dropdown/menu-item/component.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-radical/components/rad-dropdown/menu-item/component';
