+++ app/components/rad-tooltip.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-radical/components/rad-tooltip';
