+++ app/components/rad-tooltip/title.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-radical/components/rad-tooltip/title';
